requirements.txt
requests
pillow
beautifulsoup4
selenium
webdriver-manager
